Fix a bug where perf_pt tests don't receive any trace data.
For the perf_pt backend, we want to trace a thread ID (TID), not a
process (PID). A TID is a Linux specific notion, and unfortunately pid_t
is re-used for TIDs, which can lead to some confusion. I've updated doc
strings to try and make it clear to users.