Shift the tag into the high-order bits.
As suggested by Edd, this means that we avoid having to do shifting, at least
given our current layout. The current intention is that PHASE_COMPILED
contains an integer offset into a list of traces. The advantage of such an
approach is that it would allow yorickrt to have a global view of all traces
and e.g. garbage collect them. The disadvantage is that it means we might need
to garbage collect the list of traces. However, all that is too far away for us
to really know what the best solution is.

This commit also fixes a bug (also spotted by Edd) where we shifted bits one
too far to the left. This didn't cause errors, but it did mean we were
wasting a bit.