Look at the data buffer to decide what to do when the poll loop wakes.
Before we were assuming that because the poll loop woke up, it must be
due to there being data in the AUX buffer. This isn't necessarily true.
We should only look at the AUX buffer if we see a PERF_RECORD_AUX.

Now that we have a handle on PERF_RECORD_AUX samples, we can also look
inside them for indication of trace truncation.