Lift the restriction that traces can't be interrupted in tests.
Working around:
https://github.com/01org/processor-trace/issues/43

I'm hoping the outcome of this issue will be that the ptxed outcome is
the same both with and without a cache.

At some point we will need to implement a cache in our decoder anyway.

Also add a test which collects a pretty long trace. This is possible now
that we can work with interrupted traces in tests. Prior to this it was
really difficult to get a long trace with no interruption. The longer a
trace is, the more chance there is it will be interrupted (by e.g. a
context switch).