Expose PT block lengths.
We are doing this because we are having problems with MIR block
addresses matching up with PT block addresses. When we have the length
of a PT block we can check if a MIR block address is within the address
range of the PT block at least.

It's unlikely to be perfect, but it's about the best we can do with LLVM
for now.