Make the API more suitable for a genuinely multi-threaded world.
The previous API was thread-safe but not very thread-useful: there was no real
way for threads to interact with the meta-tracer. This commit rethinks things.

The main idea is that there is a single meta-tracer per process and a single
struct per thread that users (mostly) interact with. The question then becomes
how that should be structured.

The particular way this commit organises things is that we expose (stateless)
public `MT` (meta-tracer) and `MTThread` structs that feel like using objects in
a "normal" language. Each then has an Inner struct which has the state tucked
away. Users ask the public struct to do something and then it and the inner
struct can then do fine-grained locking or whatever is needed. For example, MT's
hot_threshold and active_threads are atomics, whereas MTThread's copy of
hot_threshold is non-atomic. As this may suggest, MTThread is not Sync or Send
so it can't be used by any other thread. Also, because MT and MTThread are
stateless, users are free to clone them to their heart's content: this will make
using them with move closures a lot easier, for example.

I've tried to think of one subtlety that I know we'll need going forward:
safe-points. MT/MTInner keep track of how many actively running threads there
are, so that when we come to a safe-point (which, of course, I haven't
implemented!) we can know when we're done or not. There's also another
interesting factor at play: "global" values in MT and MTThread can get out of
sync (since each each MTThreadInner has its own hot_threshold for example). I
think this is the only way to get reasonable performance, and we'll just have to
live with the consequences.

The API probably needs a lot more thought than this, but I think this starts to
nudge us in a more reasonable direction. Weirdly, the performance of the
multi-threaded benchmark isn't affected by this commit; the single-threaded
benchmark is about 50% slower. I'm not sure I'd read too much into that either
way.