The first incarnation of the pack serialiser/deserialiser.
This change adds a streamer "pack" encoder and decoder. The idea is that
compile-time info can be serialised by ykrustc at compile-time and
deserialised at run-time.

For example the MIR will need to be consulted at runtime for code
generation. We will have ykrustc serialise the MIR into packs which are
stored in the resulting ELF file ready to be used later at run-time.