Implement a couple of re-usable 128-bit numeric types for serialisation.
This enables us to use u128 and i128 in different places in the TIR
without getting bogged-down in bit-level details.

We had already implemented serialisable 128-bit TIR constants, but this
makes the implementation more generic and reusable. For example I need a
serialisable u128 in SwitchInt terminators soon.

This is necessary because of:
https://github.com/3Hren/msgpack-rust/issues/169