Flesh out the TIR API.
The idea is that if you have a SIR Trace (a collection of block
locations) and the full serialised SIR, then you can build a TIR Trace
(which will be a straight-line trace with guards).

Later in development, TIR traces will form the basis for trace
optimisation, and eventually code generation.