Fix bug when freeing decoders
Since the default decoder automatically frees up the image when
destroyed, our previous drop implementation would result in a runtime
error when default decoders are used. Using `pt_blk_free_decoder`
instead takes care of that.

Also, if a decoder fails to initialise first return the error and then
continuously return None (similar to when reaching the end of the packet
stream).

While we are here, also update the git dependencies.