Implement trace trimming.
In yktrace, tracing is turned on and off using a function calls. When we
start tracing, we trace the end of the function which starts tracing.
Similarly, when we stop tracing, we trace the start of the function
which stops tracing.

We are not interested in those parts of the trace, so we have to "trim"
the trace.

We mark the functions which start and stop tracing with `#[trace_head]`
and `#[trace_tail]` respectively. The compiler sets flags in the SIR
`Body` at compile time so that we can recognise these locations later
and do the trimming.

Notes:

Renamed `SirTraceIntoIter` to `SirTraceIterator` as the name was
confusing.

The iterator always exposes the trimmed trace. The trait functions
`loc()` and `len()` have been renamed `raw_loc()` and `raw_len()` so as
to stress that they work on the untrimmed trace.

Renamed `SIR_MAP` to just `SIR` as it now contains other things.

Updated `print_sir_trace()` debug helper to have the option to print the
trace either trimmed or raw.

There may be more than a single `#[trace_tail]` location. See the
comment in `SirMarkers`.

`ThreadTracer::stop_tracing()` now returns a `SirTrace`, not a
`TirTrace`. To make a `TirTrace` use `TirTrace::new()`. This separation
makes a few thing easier, especially testing.