Changes to SIR in preparation for Tiri.
In some places, the structure of SIR needlessly differed from MIR. It
becomes quite confusing to keep track of the two similar
representations, so we should only really diverge if we have good reason
to do so.

The representation of `Rvalue` is much more similar now, and we
introduce the notion of a `Place` into SIR, as this is an intuitive
abstraction for a storage location.