Changes for new SIR.
In short:

Sir locations are now a symbol name and a basic block index pair.

We inspect many SIR sections in the ELF binary now, not just one.

The Debug pack kind disappears entirely. We only used this to map a
DefId to a human readable name. Since we've switched to symbol names
now, that's probably already good enough.

A handful of other small changes.

For more details, see:
https://github.com/softdevteam/ykrustc/pull/80