Add a crash dump feature to aid understanding during development.
We'd like to print the assembler buffer and the register allocation at
the time of a crash. To get the buffer we need to call `finalize()` on
the assembler, but because it consumes `self` we are limited in where we
can do this.

The best way seems to be to add error handling and ripple compilation
errors up to a time where we have an owned assembler.