Rename start_vaddr and len
The current name for these variables suggests that they give the range
of the block, which would allow to calculate the next block by adding
together `start_vaddr` and `len`. This is not the case though, since len
is calculated from `block.end_ip` which returns the address of the last
instruction in this block. This means that a block with a single
instruction would have a `len` of 0, which is unintuitiv.

This commit renames the variables to `first_instr` and `last_instr` and
changes the latter to return the addresses of the last instruction of
the block (instead of returning its length).