Bring TIR traces into SSA form.
This commit turns TIR into SSA form, which means we don't have to store all
registers onto the stack before entering an inline function call
anymore.

We use alpha renaming to bring the TIR into SSA form during translation
from SIR to TIR. The idea is quite simple: whenever we see a
`StorageLive(local)` statement we assign the contained `Local` a new
variable name (`var_len`) which starts at `1` and increases with each
new variable. We store a mapping `oldname => newname` for each variable
in `rename_map`, which allows us to rename references later in the
trace.

This becomes a bit tricky when it comes to inline function calls, which
assume a clean slate, i.e. arguments are named `1..N`. Renaming those
arguments would overwrite any existing mappings, e.g. assigning `5` to
argument `1` would overwrite the mapping of a variable `1` which exists
outside of the inline function call and make it impossible to rename
references that appear after the function call to their correct new
name. Each mapping is thus a `Vector` to which we push additional
mappings. For the above example the mapping would be `1 => [2, 5]`. We
remove mappings again when they run out of scope, i.e. either when the
inlined function call ends (`Statement::Leave`), or we encounter
`Statement::StorageDead`.

The next thing we need to watch out for is that the `TraceCompiler`
assigns arguments to the correct `Local`s (which may have been renamed).
Luckily, it seems that each inline function call assigns a new variable
(to store the return value) just before entering the call. We thus can
simply assign arguments to `returnvalue+1..args.len()` instead of the
previous `1...args.len()`. However, it may turn out that not every
inlined function call assigns a new variable (though I couldn't find one
that doesn't even if the function's return type is empty). In that case
we may need to forward the "variable offset" within `Statement::Enter`.