Rename variables using offsets instead of hashmaps
The previous renaming approach was inefficient since it required us to
store vectors within a hashmap. This commit switches to a different
approach which only uses a vector. This is possible because there is no
need to rename variables in order of execution, but instead we rename
them one layer at a time, e.g. if a function `foo` calls another
function `bar`, then we first rename all variables in `foo` and
afterwards rename all variables in `bar` (technically the actual
renaming still happens in execution order, but all variables in `bar`
will have higher numbers than those in `foo`).