The beginnings of codegen for calls to external functions.
For now this allows us to call symbols with no arguments only. For now
we conservatively push all register state apart from RAX, which is
assumed to be clobbered since we are doing a function call.

In doing so, make `TraceCompiler::find_symbol` return a
Result<*mut c_void, CompileError>. This makes it more readily usable
with the `?` operator at call-sites, which are typically going to also
return `CompileError` on error.