Fix variable renaming for consecutive functions.
Currently, functions that are called consecutively (i.e. they are not
nested) use the same offset to rename their variables (the offset of
their outer context), which leads to them sharing the same variable
names. By using an accumulator which is continuously increased to
calculate the offset, we make sure consecutive functions have increasing
variable names, even when after leaving a function the offset is
temporarily reset to that of the outer context.