Rename some helper functions to be clearer.
The problem with a function name like `c_mov_int` is that it raises
questions as to where we are moving this integer. This is an attempt to
make function names like this a little easier to understand.

In the case of `c_mov_int`, I've renamed it to `mov_local_constint`,
meaning "move a constant integer to a local" (The arguments are reversed
as per Intel syntax).

Also dropping the `c_` prefix, as I think it'd be best to reserve that
convention for functions which compile some whole TIR construct, whereas
the functions I've renamed here are shared helpers called in a few
different places.