Introduce fm to ykcompile.
fm is a file matcher we can use to match the textual notation of TIR.

It's much easier to write tests this way, rather than directly matching
TIR data structures.

We only have one test that we can use fm on for now, but I'm certain we
will add more.