Fix a bug where a fixed-size buffer is easily maxed-out.
Was bitten by this on the CI server, where the filename was long. This
resulted in silent truncation of an address passed to ptxed. It took me
ages to figure out why the tests were failing!

Since this is only testing code, let's just do a heap allocation and
then we don't need to bound the argument buffer.