A more robust way of ignoring statements that use thread tracers.
We implement `used_locals()`, `maybe_defined_locals()` and
`referenced_locals()` on `Statement`, so we can check their type in
one place (just before we codegen the statement).

While we are here, we can also use this machinery for limited checking
of uses of undefined variables.

For example:
```
let x = 100;
start_tracing();
let y = x + 1
stop_tracing();
```

When building a TIR trace, we can panic when we see the use of the
undefined variable `x`. Although `x` is defined in the Rust program,
from the POV of the trace, there is no `x`. The user should have used
`trace_inputs()` to mark `x` as a trace input/output.

The approach is limited because we are unable to determine the point at
which a variable is truly live in the broader context:

For example:
```
let mut x = 1;

start_tracing();
x = 2;
stop_tracing();
```

In the trace we are unable to determine if `x=2` is a definition of a
new variable, or an update to an existing variable.

Another example:
```
struct S(u8, u8);
let mut s = S(1, 2);

start_tracing();
s.0 = 100;
stop_tracing();
```

Here we are unable to determine if `s.0 = 100` is defining `s` or if it
is mutating an existing `s`.

These issue arise because there is no explicit marker to indicate where
a variable becomes live, so we have to assume that the first write into
an as-yet undefined local defines it, but this isn't strictly true.
Ideally we'd like the above examples to panic.

Although MIR's `StorageLive` might seem like it offers a solution, we
have already determined that it isn't reliable for our purposes: MIR
locals are not consistently marked live/dead.

More on that here:
https://rust-lang.zulipchat.com/#narrow/stream/182449-t-compiler.2Fhelp/topic/Missing.20(.3F).20StorageLive.2FDead.20statements.3F

For now we are hoping that our merge point (not yet implemented) will
block the user from making use of outside variables that were not marked
as trace inputs.