Refactor projection resolution.
Refactor the code that deals with projections and prepare it for being
able to deal with chains of projections once we can query the type of a
projection.

As a nice side-effect of these changes, `Deref` has now moved out of
`Location` which reduces the branches we have to deal with when
compiling operations.

Drawback: since projections are resolved independently to the operation
they are used in, they often store the result in a temporary when it
could have been stored directly in the target location. This can likely
be optimised in the future by passing the target location to the
`resolve_projection` function.