Implement basic array indexing.
This implementation makes use of some of assumptions that are commented
by FIXMEs. Solving everything in one go would have made this PR
unneccessarily large and complicated. In particular two assumptions we
will have to deal with in the near future:

1) For now a `Deref` on values that don't fit inside a register is
always followed by another projection like `Index` or `Field`. This
means that dereferencing a copyable array/struct won't currently copy it
to the stack.

2) An `Index` projection is only applied to `&Array`. Because of 1) and
bbecause we haven't implemented array construction yet, this is a valid
assumption. Implementing indexing on `Array` means we probably have to
restructure that part of the code and use more temporaries.