Add codegen for `Place`s which don't fit it registers.
This change allows us to codegen `Place`s which get allocated to
`Location::Mem`.

In short, instead of allocating a temporary register and passing it to
`c_place()` for it to store into, `c_place()` now returns a `Location`. This
allows it to *either* create a temporary register if the `Place` fits, or
otherwise allocate it on the stack.

Then we extend `store()` to cater for all of the possible combinations that may
arise. Note though, that the combination explosion is strictly limited to
`store()`.

Also did some renaming in `store()` as it was getting confusing.

Co-authored-by: Lukas Diekmann <lukas.diekmann@gmail.com>