Add the ability to assign to non-trivial locations.
Consider the Rust snippet:

```
struct S(u8, u8, u8);
let s = S(1, 2, 3);
let sr = &mut s;
sr.1 = 6;
```

Before this change, our code-gen would do the wrong thing when
code-genning `sr.1 = 6`. It would resolve `sr.1` in isolation (without
the context that it is a left-hand side place), giving back a new
location containing the integer `2`, before overwriting it with `6`.
This new location would be distinct from `s`, so `s` wouldn't be
mutated, which isn't the required semantics.

What we need is the ability to express a reference to a place on the
left-hand side (rather than just the value itself).

To that end, there are two major changes here: 1) When resolving a
place, we can specify if that place is the left- or the right-hand-side
of an assignment. If it is the former, the resulting location will be
tagged with whether it contains a pointer to an in-memory
data-structure, which means we need an additional pointer dereference
when writing to this location. If the flag is false, we do as before. 2)
Assignments are now split into two parts: we first resolve the
right-hand-side and store it in a temporary; then we resolve the
left-hand-side and write to its location.

The change actually greatly simplifies the code-gen, as it limits the
instruction selection choices to only registers in many cases. It does
make the code-gen emit lots of unnecessary `MOV` instructions, but we
think that one day we could use a peep hole optimiser to claw back what
we lost (and probably more).

Co-authored-by: Lukas Diekmann <lukas.diekmann@gmail.com>