Statically partition the registers between locals and temps.
Currently we use the same pool of registers for local variables and
temporary registers (used for resolving projections). If we ask for a
temp register, but we have no free registers, then a local is kicked off
on to the stack.

This seemed like a good idea at the time, but it is problematic, as if
you have cached the location of a local, and in the meantime it is moved
to the stack, then you use a stale location and emit wrong code.

This change partitions the registers so that some are only for locals,
and others are only for temporaries (for now we allocate two registers
to be used for temps).

This solves the stale location problem at the cost of potentially
using more registers than is necessary. For example code which doesn't
use many projections could have benefited from those two registers that
we reserved for temps. In reality, Rust code uses projections liberally
though.

In the long run, we should decompose projection chains to simpler
assignment statements, thus removing the need for temporary registers at
all.