New IR.
This is a new design for SIR (and therefore TIR). The new design
resolves as many projections at SIR generation time so that they don't
have to be done at runtime. By doing so, we lessen the burden on the
trace code generator and the register allocator.

Before, in addition to a regular allocator, we also had an allocator for
temporary registers. This led to annoying situations where many operations
needed temporary storage space, but registers were full and locals had to be
evicted to the stack. This led to unnecessary complexity, and further you had
to free temporaries at the right point to avoid leaks.

In this new IR, we simply have one register reserved as a temporary. This is
possible because there is reduced need for temporaries. We've tried hard to
make everything that requires storage space to be a proper local variable
handled by the regular register allocator.

Temporaries are still needed for certain operations (e.g. as an intermediary
for mem-to-mem stores), but their scope is strictly very short and we only ever
need one temporary register.

Co-authored-by: Lukas Diekmann <lukas.diekmann@gmail.com>