Add early version of SIR interpreter.
The SIR interpreter runs on SIR bodies and will allow us to fill the gap
between running a compiled trace, and running the user interpreter, and
provides a way to safely get back to the latter.

When a guard inside a compiled trace fails, we want return to the user
interpreter. However, we can't simply jump back into the user
interpreter as we might be in the middle of a bytecode instruction. The
SIR interpreter runs on SIR instructions, and is initialised to continue
interpreting as the exact point where the guard failed (the guard will
make sure that the SIR interpreters frames are initialised accordingly).
The SIR interpreter is then run until it reaches the next control point,
which is typically at the beginning of the next bytecode, at which point
the user interpreter kicks back in. In RPython this technique is called
blackholing.

The SIR interpreter runs on SIR instructions, which are meant to be
compiled to machine code. It thus has to deal with instructions which
often use pointer arithmetic to address specific offsets within a Local.
The SIR interpreter thus allocates memory to store the Locals and read
and write to them.