Implement lazy SIR loading.
Currently we load all of the SIR into memory the first time we request a
type or a body. By doing so we incur a large (one-time) cost, even if we
don't use 99% of the loaded SIR.

This change implements lazy loading. Each SIR section now starts with a
header which serves as a table of contents, mapping type and body
identifiers to their byte offsets within the section. This gives us
random access to the contents of the SIR, which is both faster and more
memory efficient.

This first version is naive in that it implements no caching (of types,
bodies, or any of the associated decoding mechanisms). It can surely be
optimised further (but later).

Rough benchmarking of a full test suite run before and after this
change:

              Mean        Std.Dev.    Min         Median      Max
Before:
  real        30.643      0.325       30.246      30.638      31.245
  user        127.624     0.931       126.417     127.238     129.612
  sys         5.106       0.680       3.925       5.270       6.062
After:
  real        22.399      0.614       21.910      22.159      24.056
  user        131.255     0.653       129.918     131.371     132.273
  sys         4.981       0.742       4.145       4.953       6.603

The next-most-costly part of the system is now dwarf stuff.