Remove the `Arc` from inside `Location`.
`Arc`s are bulky -- they malloc memory and, on top of whatever overhead
the malloc library imposes, store two extra machine words. That indirection is
also going to induce some slowdown (though that's harder to quantify).

We want `Location`s to be small and fast. This commit makes a `Location` just a
`AtomicUsize`, which is as small and fast as I think we can possibly make it.

However, I assume one of the reasons we moved to using `Arc` is that it gives us
a way of saying "is the current thread tracing the current Location?" using Arc
pointer equality. Moving to just an AtomicUsize removes that possibility from
us.

The problem then becomes "how do I know if this Location is the one I was
tracing?" We don't want to rely on the address of the Location, since that
imposes a burden on the user to place it somewhere where it can't move (which
they could easily forget, leading to all sorts of pain).

I have therefore borrowed a cunning trick from RPython, whereby whilst -- and
only whilst -- we're `PHASE_TRACING` a Location, we malloc a byte of memory and
use that as a unique identifier for the Location. Even if the `Location moves
around, the malloc'd address won't move, so we can always identify the
`Location`. We only pay this cost for `Location`s that we're tracing, so it's
fairly cheap.

In order that we don't leak memory, or double free, we can no loner allow
`Location`s to be `Clone`able. I think this is probably a good thing, as it
makes clearer to the user the idea that "`Location`s are sort-of unique." It
does mean tweaking some tests so that `Location`s are stored in a way that
multiple threads can access. In this case, amusingly, the easiest way to solve
the problem is to put `Location`s in tests inside `Arc`s. *However*, there's no
reason why a real, high performance interpreter would do the same: I think
they'd just embed a `Location` (which is just a `usize` after all) directly in
their bytecode.