Fix two threading bugs.
Previously there were two flaws in the Location statemachine, one fundamental,
and one based on not reading the Rust docs correctly.

The fundamental flaw relates to the fact that when we move from PHASE_COUNTING
to PHASE_COMPILING we have to allocate temporary storage to store the compiled
trace in, which PHASE_COMPILING must free when it moves to PHASE_COMPILED.
Previously there was a race condition where two threads could observe
PHASE_COMPILING as having completed and both would try freeing the temporary
storage leading to either a double free being detected or a segfault (I observed
both happening!). This commit fixes that by introducing a new state into the
state machine PHASE_LOCKED: when we're in PHASE_COMPILING and we want to see if
the trace has been compiled, we temporarily flip the state to PHASE_LOCKED, at
which point other threads know to leave the Location alone. If the trace has
been compiled, we can free the temporary storage without worrying that we're
racing to another thread; if it hasn't been compiled we go back to
PHASE_COMPILING. The statemachine thus now looks as follows:

               ┌────────────────────────────────────────┐
               │                                        │ ─────────────┐
   reprofile   │             PHASE_COUNTING             │              │
  ┌──────────▶ │                                        │ ◀────────────┘
  │            └────────────────────────────────────────┘    increment
  │              │                                 ▲         count
  │              │ start tracing                   │
  │              ▼                                 │
  │            ┌────────────────────┐              │ abort
  │      ┌─────│                    │              |
  │      |     │   PHASE_TRACING    │ ─────────────┤
  │      └────▶│                    │              |
  │            └────────────────────┘              │
  │              │ start compiling trace           │
  │              │ in thread                       │
  │              ▼                                 │
  │            ┌────────────────────┐              │
  │            │                    │              |
  |            |  PHASE_COMPILING   | ─────────────┤
  │            │                    │              |
  │            └────────────────────┘              |
  │              │ trace maybe    ▲                |
  │              │ compiled       | trace not yet  |
  │              ▼                | compiled       |
  │            ┌────────────────────┐              |
  │            │    PHASE_LOCKED    │ ─────────────┘
  │            └────────────────────┘
  │              │ trace definitely
  │              │ compiled
  │              ▼
  │            ┌────────────────────┐
  │            │                    │ ──────┐
  │            │   PHASE_COMPILED   │       │
  └─────────── │                    │ ◀─────┘
               └────────────────────┘

The second flaw relates to the fact that we used Rust channels to pass a
compiled trace from the compiling thread to an interpreter thread, meaning
that multiple threads could simultaneously try reading from the channel.
However, while the Receiver type is Send, it is not Sync, so this was not safe:
we didn't notice this because we converted the Reciever into a box and passed a
tagged pointer around, bypassing Rust's type system. This commit fixes this
problem by using an explicit Mutex, which is safe to pass between threads via
our tagged pointers.

The new test "simple_multithreaded_interpreter" starts multiple threads on a
small number of locations, with different threads non-deterministically moving
amongst the Location statemachine. Because it's inherently non-deterministic, on
any given run it doesn't necessarily trigger the problems fixed above, but if
run often enough, it is capable of hitting both. We thus run it several times
with the aim that neither of the above problems can reappear in the future.

Note that while this commit aims to solve crashing bugs, it now explicitly shows
that we have two potential memory leaks, both related to the fact that we can't
guarantee that any given Location will ever execute again. At some point we will
need to add some sort of timeout that frees traces that haven't recently been
executed. We also don't deal with the situation where a loop starts tracing but
doesn't actually loop: the Location in question currently ends up stuck in
PHASE_TRACING forever. Fixing these things is for the future.